Update testing.yml
This is kind of dirty, but locally you need to run the application to load the data, then run the tests once the data is loaded. This will obviously change in the future, but until then, this gets this to a more stable state in the meantime. 

Basically the end goal would be to couple the docker database and the execution script together. Then update this workflow so that it uses the same set up as the docker image (which at this current time it does) and then update this from running the application to running the isolated program to load the data. 

Pretty simple change once it's handled.